requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
pydantic>=2.8.0
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6
datadog==0.50.0

# Linkup Web Scraping Dependencies
linkup-sdk==0.6.0
openai==1.3.0
